Move benefits to data folder
